backend/requirements.txt
beautifulsoup4
requests
python-dotenv
Flask